console: Move message data formatting to types util
